+++ R/align-panel.R
@@ -9,7 +9,7 @@
 #' @inheritParams ggplot2::ggplot
 #'
 #' @section ggplot2 specification:
-#' `align_panel` initializes a `ggplot` data and `mapping`.
+#' `align_panel` initializes a ggplot `data` and `mapping`.
 #'
 #' `align_panel()` always applies a default mapping for the axis of the data
 #' index in the layout. This mapping is `aes(y = .data$.y)` for horizontal stack