+++ R/alignpatch-guides.R
@@ -1,10 +1,9 @@
-#' @importFrom utils getFromNamespace
 #' @importFrom rlang try_fetch
 assemble_guides <- function(guides, guide_pos, theme) {
     # for every position, collect all individual guides and arrange them
     # into a guide box which will be inserted into the main gtable
     package_box <- try_fetch(
-        .subset2(getFromNamespace("Guides", "ggplot2"), "package_box"),
+        .subset2(ggfun("Guides"), "package_box"),
         error = function(cnd) package_box
     )
     package_box(guides, guide_pos, theme)